% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePower.R
\name{estimatePower}
\alias{estimatePower}
\title{Estimate Power of the Actual Data}
\usage{
estimatePower(inputObject, groupVec, isLogTransformed = FALSE,
  dataType = c("RNASeq", "Proteomics"), minLFC = 0.5, alpha = 0.05,
  ST = 100, seed = 123, showSimProcess = FALSE, saveResultData = FALSE)
}
\arguments{
\item{inputObject}{a numeric raw data matrix or SummarizedExperiment object}

\item{groupVec}{a vector indicating the grouping of samples}

\item{isLogTransformed}{logical; if \code{TRUE},
input data will be used directly without further transformation}

\item{dataType}{"RNASeq" or "Proteomics"
indictes the data type of the input data matrix.}

\item{minLFC}{the threshold for log2 fold change,
entrys with lower LFC are not included in the power calculation,
set to 0 if no threshold is needed.}

\item{alpha}{controlled false positive rate.}

\item{ST}{the number of simulations of abundance data generation and
repeated times of statistical test for each entry (>=100 recommended).}

\item{seed}{an integer seed for the random number generator.}

\item{showSimProcess}{logical; if \code{TRUE}, show the detailed 
information of each simulation, used for debugging only.}

\item{saveResultData}{logical; if \code{TRUE},
save the simulated data into RData with name pattern
"simulated_Data_numRep_X_numSim_XXX_XXXXX.RData"}
}
\value{
a list of power estimates grouped in comparisons
between each two groups
}
\description{
Estimate power of comparison between each two groups based on
the data simulated from estimated normal distributions of entrys
in the entire dataset
}
\examples{
# Example 1: a random generated Proteomics dataset (10 DE, 100 non-DE)
# Note: Simulation times(ST) is specified as 10 for shorter example runtime
#       For better performence, ST > 50 is recommended
data(exampleProteomicsData)
dataMatrix <- exampleProteomicsData$dataMatrix
groupVec <- exampleProteomicsData$groupVec

# Run estimation without LFC filtration
resObject <- estimatePower(dataMatrix, groupVec,
                           dataType="Proteomics",
                           isLogTransformed=FALSE,
                           minLFC=0, alpha=0.05, 
                           ST=10, seed=123)

# Run estimation with minLFC=1 (Fold Change=2)
resObject2 <- estimatePower(dataMatrix, groupVec,
                           dataType="Proteomics",
                           isLogTransformed=FALSE,
                           minLFC=1, alpha=0.05, 
                           ST=10, seed=123)

# Example 2: a random generated RNASeq dataset (10 DE, 100 non-DE)
# Note: Simulation times(ST) is specified as 10 for shorter example runtime
#       For better performence, ST > 50 is recommended
data(exampleRNASeqData)
dataMatrix <- exampleRNASeqData$dataMatrix
groupVec <- exampleRNASeqData$groupVec

# Run estimation without LFC filtration
resObject <- estimatePower(dataMatrix, groupVec,
                             dataType="RNASeq",
                             isLogTransformed=FALSE,
                             minLFC=0, alpha=0.05, 
                             ST=10, seed=123)

# Run estimation with minLFC=1 (Fold Change=2)
resObject2 <- estimatePower(dataMatrix, groupVec,
                            dataType="RNASeq",
                            isLogTransformed=FALSE,
                            minLFC=1, alpha=0.05, 
                            ST=10, seed=123)


}
\seealso{
\code{\link{predictPower}}
predict power with incresing sample sizes
}
