% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleEstimatedPower.R
\docType{data}
\name{exampleEstimatedPower}
\alias{exampleEstimatedPower}
\title{An Example Power Object For Plot Test}
\format{An object of class \code{list} of length 10.}
\usage{
data(exampleEstimatedPower)
}
\description{
This is an example estimated power object calculated
from a spiked-in proteomics dataset with 5 groups.
}
\examples{

data(exampleEstimatedPower)
plotCurrentPower(powerList = exampleEstimatedPower,
                 savePlot = FALSE)

}
\keyword{datasets}
