% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePredictedPower.R
\docType{data}
\name{examplePredictedPower}
\alias{examplePredictedPower}
\title{An Example Predicted Power Object For Plot Tests}
\format{An object of class \code{array} of dimension 2794 x 5 x 3.}
\usage{
data(examplePredictedPower)
}
\description{
This is an example predicted power object calculated from
a spiked-in proteomics dataset, the predictions are performed under
replicate number 10, 15, 20, 25, and 30.
}
\examples{
data(examplePredictedPower)
plotPredictedPower(examplePredictedPower,
                  plotType="lineplot",
                  minLFC=0, maxLFC=4,
                  LFCscale=0.5)

}
\keyword{datasets}
