% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSimulatedData.R
\docType{data}
\name{exampleSimulatedData}
\alias{exampleSimulatedData}
\title{Example Simulated Proteomics Data}
\format{An object of class \code{list} of length 2.}
\usage{
data(exampleSimulatedData)
}
\description{
This is an example simulated data object
used for demostrate function \code{plotStatisticsDensity}
}
\examples{

data(exampleSimulatedData)

}
\keyword{datasets}
