% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMapPredPwr.R
\docType{methods}
\name{heatMapPredPwr}
\alias{heatMapPredPwr}
\alias{heatMapPredPwr,PEObject-method}
\title{Observe Predicted Power Within Different LFC Scales.}
\usage{
heatMapPredPwr(PEObject, minLFC, maxLFC, LFCscale)

\S4method{heatMapPredPwr}{PEObject}(PEObject, minLFC, maxLFC, LFCscale)
}
\arguments{
\item{PEObject}{the result object \code{PEObject} returned from
\code{predictPower}.}

\item{minLFC}{default as 0, the left edge of the LFC range within
which genes will be included in the graph.}

\item{maxLFC}{default as NA (the maximun value in data will be used) the
right edge of the LFC range within which genes will be included in the graph.}

\item{LFCscale}{the size of each unit when segmenting predicted power by LFC}
}
\value{
plot(s) of power density under multiple sample sizes
}
\description{
With a complete power list and LFC list returned from
\code{predictPower}, power estimates can be observed dynamically
within specified LFC ranges.
}
\examples{
# load an example onject containing
# predicted power under different sample sizes
data(examplePEObject)
# plot a heatmap
heatMapPredPwr(examplePEObject,
               LFCscale=1)
heatMapPredPwr(examplePEObject,
               LFCscale=1)
#It is possible to observe power trend in different scales and ranges of LFCs
heatMapPredPwr(examplePEObject,
            minLFC=0,
            maxLFC=3,
            LFCscale=0.5)
heatMapPredPwr(examplePEObject,
            minLFC=0,
            maxLFC=3,
            LFCscale=1)
}
