% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPower.R
\name{listEstPwr}
\alias{listEstPwr}
\title{List Estimated Power}
\usage{
listEstPwr(PEObject, decreasing = TRUE, top = 20, selected = NA)
}
\arguments{
\item{PEObject}{the input PEObject.}

\item{decreasing}{logical; TRUE, decreasing order; FALSE, increasing order.}

\item{top}{the number of genes/proteins in the top list}

\item{selected}{default as NA; specify as a list of geneID or protein ID
to show power of a list of interested records.}
}
\value{
a top list of power / power of a list of interested genes or proteins
}
\description{
show a top-list of power in numerical order, 
or list the power selected genes/proteins.
}
\examples{
data(examplePEObject)
# show 10 top proteins with high power (decreasing order)
listEstPwr(examplePEObject, decreasing = TRUE, top = 10)
# show a list of interested proteins
listEstPwr(examplePEObject, selected = c("Protein_1", "Protein_11"))

}
