% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPower.R
\name{listPredPwr}
\alias{listPredPwr}
\title{List Predicted Power}
\usage{
listPredPwr(inputObject, decreasing = TRUE, top = 20, selected = NA)
}
\arguments{
\item{inputObject}{the input inputObject.}

\item{decreasing}{logical; TRUE, decreasing order;
FALSE, increasing order.}

\item{top}{the number of genes/proteins in the top list}

\item{selected}{default as NA;
specify as a list of geneID or protein ID
to show power of a list of interested records.}
}
\value{
a top list of power / power of a list of interested
genes or proteins for each sample size
}
\description{
show a top-list of predicted power in numerical order,
or list the power selected genes/proteins.
}
\examples{
data(exampleObject)
# show 10 top proteins with high power (decreasing order)
listPredPwr(exampleObject, decreasing = TRUE, top = 10)

# show a list of interested proteins
listPredPwr(exampleObject, selected = c("Protein_1", "Protein_11"))

}
