% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEObject.R
\docType{class}
\name{PEObject-class}
\alias{PEObject-class}
\alias{PEObject}
\title{PowerExplorer Object}
\description{
An extended \code{SummarizedExpriment} object to 
contain input dataMatrix, grouping information, estimated power, 
predicted power, fold change estimates and other estimation
parameters.
}
