% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurrentPower.R
\name{plotCurrentPower}
\alias{plotCurrentPower}
\title{Plot A Summary of Estimated Power}
\usage{
plotCurrentPower(powerList, savePlot = FALSE)
}
\arguments{
\item{powerList}{the \code{power.list}
returned from \code{estimateCurrentPower}.}

\item{savePlot}{logical; if \code{TRUE},
save the plot into folder "savedPlots" in work directory.}
}
\value{
plot(s) of the summarised information on the estimated power
}
\description{
Produce a plot to summary the power estimated by function
\code{estimateCurrentPower}, the plot function is called in
\code{estimateCurrentPower}, but using it manually is possible
}
\examples{
data(exampleEstimatedPower)
plotCurrentPower(exampleEstimatedPower, savePlot=TRUE)

}
