% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimatedPower.R
\name{plotEstimatedPower}
\alias{plotEstimatedPower}
\title{Plot A Summary of Estimated Power}
\usage{
plotEstimatedPower(powerList)
}
\arguments{
\item{powerList}{the \code{power.list}
returned from \code{estimateCurrentPower}.}
}
\value{
plot(s) of the summarised information on the estimated power
}
\description{
Produce a plot to summary the power estimated by function
\code{estimateCurrentPower}, the plot function is called in
\code{estimateCurrentPower}, but using it manually is possible
}
\examples{
data(exampleEstimatedPower)
plotEstimatedPower(exampleEstimatedPower)

}
