% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstPwr.R
\name{plotEstPwr}
\alias{plotEstPwr}
\title{Plot A Summary of Estimated Power}
\usage{
plotEstPwr(inputObject)
}
\arguments{
\item{inputObject}{a result container object 
\code{\link{PowerExplorerStorage}} returned from 
\code{\link{estimatePower}}.}
}
\value{
plot(s) of the summarised information on the estimated power
}
\description{
Produce a plot to summary the power estimated by function
\code{\link{estimatePower}}, the plot function is called in
\code{\link{estimatePower}}, but using it manually is possible
}
\examples{
data(exampleObject)
plotEstPwr(exampleObject)

}
