% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictedPower.R
\name{plotPredictedPower}
\alias{plotPredictedPower}
\title{Observe Predicted Power Within Different LFC Scales}
\usage{
plotPredictedPower(predictedPower, plotType = c("lineplot", "heatmap"),
  minLFC = 0, maxLFC = NA, LFCscale = NA)
}
\arguments{
\item{predictedPower}{the \code{ predictedPower} returned from
\code{predictSampleSizePower}.}

\item{plotType}{choose graph type between lineplot and heatmap to
observe the predicted power from different aspects}

\item{minLFC}{default as 0, the left edge of the LFC range within
which genes will be included in the graph.}

\item{maxLFC}{default as NA (the maximun value in data will be used) the
right edge of the LFC range within which genes will be included in the graph.}

\item{LFCscale}{the size of each unit when segmenting predicted power by LFC}
}
\value{
plot(s) of power density under multiple sample sizes
}
\description{
With a complete power list and LFC list returned from
\code{predictSampleSizePower}, power estimates can be observed dynamically
within specified LFC ranges.
}
\examples{
# load an example onject containing
# predicted power under different sample sizes
data(examplePredictedPower)
# plot a heatmap
plotPredictedPower(examplePredictedPower,
                   plotType="heatmap",
                   LFCscale=1)
plotPredictedPower(examplePredictedPower,
                   plotType="lineplot",
                   LFCscale=1)
#It is possible to observe power trend in different scales and ranges of LFCs
plotPredictedPower(examplePredictedPower,
                   plotType="lineplot",
                   minLFC=0,
                   maxLFC=3,
                   LFCscale=0.5)
plotPredictedPower(examplePredictedPower,
                   plotType="heatmap",
                   minLFC=0,
                   maxLFC=3,
                   LFCscale=1)
}
