% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStatisticsDensity.R
\name{plotStatisticsDensity}
\alias{plotStatisticsDensity}
\title{Plot t-Statistics Density Between Null and Alternative Hypothesis}
\usage{
plotStatisticsDensity(simulatedData, entryID, alpha = 0.05)
}
\arguments{
\item{simulatedData}{the simulated data object}

\item{entryID}{the gene name (string) or the row number (numeric)
in the countData oject}

\item{alpha}{controlled false positive rate}
}
\value{
A plot presents the t-statistics density between null and
alternative hypothesis
}
\description{
With saved simulated data, the t statistics under both null and
alternative hypotheses of a single protein can be observed by giving the
protein ID name.
}
\examples{
#if the simulated data were saved
data("exampleSimulatedData")
plotStatisticsDensity(simulatedData=exampleSimulatedData,
                     entryID="A3KMH1",
                     alpha=0.05,
                     savePlot=TRUE)
}
