% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerExplorerStorage-class.R
\docType{class}
\name{PowerExplorerStorage-class}
\alias{PowerExplorerStorage-class}
\alias{PowerExplorerStorage}
\title{PowerExplorer Object}
\description{
An extended \code{SummarizedExpriment} object to 
contain input dataMatrix, grouping information, estimated power, 
predicted power, fold change estimates and other estimation
parameters.
}
