% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSampleSizePower.R
\name{predictSampleSizePower}
\alias{predictSampleSizePower}
\title{Estimate Power Under Increasing Sample Sizes}
\usage{
predictSampleSizePower(inputDataMatrix, groupVec, isLogTransformed = FALSE,
  dataType = c("RNA-Seq", "Proteomics"), rangeSimNumRep = NA,
  alpha = 0.05, ST = 100, seed = 123, showProcess = FALSE,
  saveSimulatedData = TRUE)
}
\arguments{
\item{inputDataMatrix}{a numeric raw Proteomics abundance data matrix,
in which rows correspond to proteins and columns correspond to samples.}

\item{groupVec}{a vector indicating the grouping of samples}

\item{isLogTransformed}{logical; if \code{TRUE}, input data will be used
directly without further transformation}

\item{dataType}{"RNA-Seq" or "Proteomics" indictes the data type of
the input data matrix.}

\item{rangeSimNumRep}{a vector of sample sizes under which power
will be estimated}

\item{alpha}{controlled false positive rate.}

\item{ST}{the number of simulations of abundance data generation and
repeated times of statistical test for each protein (>=100 recommended).}

\item{seed}{an integer seed for the random number generator.}

\item{showProcess}{logical; if \code{TRUE}, show the detailed information of
each simulation, used for debugging only.}

\item{saveSimulatedData}{logical; if \code{TRUE}, save the simulated data
into RData with name pattern
"simulated_Data_numRep_X_numSim_XXX_XXXXX.RData"}
}
\value{
a list of power predictions for each sample size, grouped in
comparisons between each two groups
}
\description{
Simlilar to \code{\link{estimateCurrentPower}},
power estimations are performed under multiple increasing sample sizes
}
\examples{
# Example 1: a random generated Proteomics dataset (10 DE, 100 non-DE)
data(exampleProteomicsData)
dataMatrix <- exampleProteomicsData$dataMatrix
groupVec <- exampleProteomicsData$groupVec

# Run estimation
# Note: Simulation times(ST) is specified as 5 for shorter example runtime
#       For better performence, ST > 50 is recommended
predictedPower <- predictSampleSizePower(dataMatrix, groupVec,
                                         isLogTransformed=FALSE,
                                         dataType="Proteomics",
                                         rangeSimNumRep=c(5, 10, 15),
                                         alpha=0.05, ST=5, seed=123,
                                         showProcess=FALSE,
                                         saveSimulatedData=TRUE)
# Example 2: a random generated RNA-Seq dataset (10 DE, 100 non-DE)
data(exampleRNASeqData)
dataMatrix <- exampleRNASeqData$dataMatrix
groupVec <- exampleRNASeqData$groupVec

# Run estimation
# Note: Simulation times(ST) is specified as 5 for shorter example runtime
#       For better performence, ST > 50 is recommended
predictedPower <- predictSampleSizePower(dataMatrix, groupVec,
                                         isLogTransformed=FALSE,
                                         dataType="RNA-Seq",
                                         rangeSimNumRep=c(5, 10, 15),
                                         alpha=0.05, ST=5, seed=123,
                                         showProcess=FALSE,
                                         saveSimulatedData=TRUE)
}
\seealso{
\code{\link{estimateCurrentPower}} estimate power based on
actual data
}
