% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerExplorerStorage-class.R
\docType{methods}
\name{show,PowerExplorerStorage-method}
\alias{show,PowerExplorerStorage-method}
\title{show method for PowerExplorerStorage}
\usage{
\S4method{show}{PowerExplorerStorage}(object)
}
\arguments{
\item{object}{a PowerExplorerStorage object as input}
}
\value{
a summary of input PowerExplorerStorage object
}
\description{
show method for PowerExplorerStorage
}
\section{Methods (by class)}{
\itemize{
\item \code{PowerExplorerStorage}: method for PowerExplorerStorage objects
}}

\examples{
data(exampleObject)
show(exampleObject)
}
