% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePEObject.R
\docType{data}
\name{examplePEObject}
\alias{examplePEObject}
\title{An Example Predicted Power Object For Plot Tests}
\format{An object of class \code{PEObject} with 110 rows and 15 columns.}
\usage{
data(examplePEObject)
}
\description{
This is an example PEObject
}
\examples{
data(examplePEObject)

show(examplePEObject)
plotEstPwr(examplePEObject)
}
\keyword{datasets}
