% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlotPredPwr.R
\name{linePlotPredPwr}
\alias{linePlotPredPwr}
\title{Observe Predicted Power Within Different LFC Scales}
\usage{
linePlotPredPwr(PEObject, minLFC, maxLFC, LFCscale)
}
\arguments{
\item{PEObject}{the result object \code{PEObject} returned from
\code{predictPower}.}

\item{minLFC}{default as 0, the left edge of the LFC range within
which genes will be included in the graph.}

\item{maxLFC}{default as NA (the maximun value in data will be used) the
right edge of the LFC range within which genes will be included in the graph.}

\item{LFCscale}{the size of each unit when segmenting predicted power by LFC}
}
\value{
plot(s) of power density under multiple sample sizes
}
\description{
With a complete power list and LFC list returned from
\code{predictPower}, power estimates can be observed dynamically
within specified LFC ranges.
}
\examples{
# load an example onject containing
# predicted power under different sample sizes
data(examplePredictedPower)
# plot a heatmap
plotPredPwr(examplePredictedPower,
                   LFCscale=1)
plotPredPwr(examplePredictedPower,
                   LFCscale=1)
#It is possible to observe power trend in different scales and ranges of LFCs
plotPredPwr(examplePredictedPower,
                   minLFC=0,
                   maxLFC=3,
                   LFCscale=0.5)
plotPredPwr(examplePredictedPower,
                   minLFC=0,
                   maxLFC=3,
                   LFCscale=1)
}
