% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstPwr.R
\name{plotEstPwr}
\alias{plotEstPwr}
\title{Plot A Summary of Estimated Power}
\usage{
plotEstPwr(PEObject)
}
\arguments{
\item{estimatedPower}{the \code{power.list}
returned from \code{estimatePower}.}
}
\value{
plot(s) of the summarised information on the estimated power
}
\description{
Produce a plot to summary the power estimated by function
\code{estimatePower}, the plot function is called in
\code{estimatePower}, but using it manually is possible
}
\examples{
data(exampleestimatedPowerer)
plotEstPwr(exampleestimatedPowerer)

}
