% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEObject.R
\docType{methods}
\name{show,PEObject-method}
\alias{show,PEObject-method}
\title{show method for PEObject}
\usage{
\S4method{show}{PEObject}(object)
}
\arguments{
\item{object}{a PEObject object as input}
}
\value{
a summary of input PEObject object
}
\description{
show method for PEObject
}
\section{Methods (by class)}{
\itemize{
\item \code{PEObject}: method for PEObject objects
}}

\examples{
data(examplePEObject)
show(examplePEObject)
}
